+++ extensions/measurement-tracking/src/viewports/TrackedCornerstoneViewport.tsx
@@ -140,17 +140,18 @@ function TrackedCornerstoneViewport(props) {
           // Only send the tracked measurements event for the active viewport to avoid
           // sending it more than once.
           if (viewportId === activeViewportId) {
-            const { referenceStudyUID: StudyInstanceUID,
+            const {
+              referenceStudyUID: StudyInstanceUID,
               referenceSeriesUID: SeriesInstanceUID,
-              uid: measurementId } =
-              measurement;
+              uid: measurementId,
+            } = measurement;
 
             sendTrackedMeasurementsEvent('SET_DIRTY', { SeriesInstanceUID });
             sendTrackedMeasurementsEvent('TRACK_SERIES', {
               viewportId,
               StudyInstanceUID,
               SeriesInstanceUID,
-              measurementId
+              measurementId,
             });
           }
         }).unsubscribe