+++ platform/core/src/index.test.js
@@ -1,7 +1,7 @@
 import * as OHIF from './index';
 
 describe('Top level exports', () => {
-  test('have not changed', () => {
+  test.only('have not changed', () => {
     const expectedExports = [
       'MODULE_TYPES',
       //
@@ -46,6 +46,7 @@ describe('Top level exports', () => {
       'pubSubServiceInterface',
       'PubSubService',
       'PanelService',
+      'WorkflowStepsService',
       'useToolbar',
     ].sort();