+++ platform/ui/src/index.js
@@ -31,6 +31,7 @@ export {
 
 /** COMPONENTS */
 export {
+  ActionButtons,
   AboutModal,
   AdvancedToolbox,
   AllInOneMenu,
@@ -50,6 +51,7 @@ export {
   EmptyStudies,
   ErrorBoundary,
   ExpandableToolbarButton,
+  LineChart,
   ListMenu,
   Icon,
   IconButton,
@@ -84,6 +86,7 @@ export {
   SegmentationGroupTableExpanded,
   SidePanel,
   SplitButton,
+  ProgressDropdown,
   LegacySplitButton,
   StudyBrowser,
   StudyItem,
@@ -118,6 +121,7 @@ export {
   ViewportDownloadForm,
   ViewportGrid,
   ViewportPane,
+  WindowLevel,
   WindowLevelMenuItem,
   ImageScrollbar,
   ViewportOverlay,